+++ src/bundler/gem-dependency.ts
