+++ src/bundler/stub-specification.ts
