+++ src/definition/a-disposable.ts
