+++ src/definition/a-view-emitter.ts
@@ -1,6 +1,6 @@
 import { EventEmitter, Event } from 'vscode';
-import { IEntry } from './i_entry';
-import { ADisposable } from './a_disposable';
+import { IEntry } from './i-entry';
+import { ADisposable } from './a-disposable';
 
 export abstract class ViewEmitter extends ADisposable {
   protected aEntryEmitter: EventEmitter<IEntry | undefined> = new EventEmitter<IEntry | undefined>();
