+++ src/definition/i-entry.ts
