+++ src/view/general/general-entry.ts
@@ -1,12 +1,12 @@
 import * as _ from 'lodash';
 import { basename, dirname, join as pjoin } from 'path';
 import { FileType, TreeItem, Uri } from 'vscode';
-import { IEntry } from '../../definition/i_entry';
+import { IEntry } from '../../definition/i-entry';
 import { Spec } from '../../spec';
 import { Utils } from '../../util';
 import { FileStat } from '../../lib/file-stat';
-import { SpecEntry } from '../spec/spec_entry';
-import { GeneralItem } from './general_item';
+import { SpecEntry } from '../spec/spec-entry';
+import { GeneralItem } from './general-item';
 
 export class GeneralEntry implements IEntry {
   constructor(
