+++ src/view/spec/spec-item.ts
